+++ dev_workbench/lib/src/scenario.dart
@@ -1,15 +1,18 @@
-import 'package:dev_workbench/src/scenarios/scenario_config.dart';
-import 'package:dev_workbench/src/scenarios/scenario_configurator.dart';
-import 'package:dev_workbench/src/scenarios/scenario_screen.dart';
-import 'package:dev_workbench/src/scenarios/tab_header_row/tab_header_row_scenario.dart';
-import 'package:dev_workbench/src/scenarios/tab_label_builder/tab_label_builder.dart';
-import 'package:dev_workbench/src/scenarios/theme/theme_config.dart';
-import 'package:dev_workbench/src/scenarios/theme/theme_configurator.dart';
-import 'package:dev_workbench/src/scenarios/theme/theme_scenario.dart';
+import 'package:dev_workbench/src/scenario_configurator.dart';
+import 'package:dev_workbench/src/scenario_screen.dart';
+import 'package:dev_workbench/src/scenarios/tab_header_row_scenario.dart';
+import 'package:dev_workbench/src/scenarios/tab_label_builder_scenario.dart';
+import 'package:dev_workbench/src/scenarios/theme_scenario.dart';
+import 'package:flutter/material.dart';
+
+import 'scenarios/tab_remove_interceptor_scenario.dart';
+import 'scenarios/tab_secondary_tap_scenario.dart';
 
 enum Scenario {
   theme('Theme', ThemeScenario.builder),
-  tabLabelBuilder('TabLabelBuilder', TabLabelBuilder.builder),
+  tabLabelBuilder('TabLabelBuilder', TabLabelBuilderScenario.builder),
+  tabSecondaryTap('Tab secondary tap', TabSecondaryTapScenario.builder),
+  tab('Tab remove interceptor', TabRemoveInterceptorScenario.builder),
   tabHeaderRow('TabHeaderRow', TabHeaderRowScenario.builder);
 
   const Scenario(this.text, this.builder);
@@ -18,4 +21,7 @@ enum Scenario {
   final ScenarioBuilder builder;
 }
 
-typedef ScenarioBuilder = (ScenarioConfigurator?, ScenarioScreen) Function();
+typedef ScenarioBuilder =
+    (ScenarioConfigurator?, ScenarioScreen) Function({
+      required GlobalKey<ScaffoldMessengerState> scaffoldMessengerKey,
+    });