+++ dev_workbench/lib/src/scenario_configurator.dart
@@ -1,15 +1,38 @@
-import 'package:dev_workbench/src/scenarios/scenario_configurator.dart';
-import 'package:dev_workbench/src/scenarios/theme/theme_config.dart';
+import 'package:dev_workbench/src/scenario_config.dart';
 import 'package:flutter/material.dart';
 
-import '../../widgets/brightness_chooser.dart';
-import '../../widgets/position_chooser.dart';
-import '../../widgets/selector.dart';
-import '../../widgets/side_tabs_layout_chooser.dart';
-import '../../widgets/theme_chooser.dart';
+import 'widgets/brightness_chooser.dart';
+import 'widgets/position_chooser.dart';
+import 'widgets/selector.dart';
+import 'widgets/side_tabs_layout_chooser.dart';
+import 'widgets/theme_chooser.dart';
 
-class ThemeConfigurator extends ScenarioConfigurator<ThemeConfig> {
-  const ThemeConfigurator({super.key, required super.config});
+abstract class ScenarioConfigurator<CONFIG extends ScenarioConfig>
+    extends StatelessWidget {
+  const ScenarioConfigurator({super.key, required this.config});
+
+  final CONFIG config;
+
+  @override
+  Widget build(BuildContext context) {
+    return ListenableBuilder(
+      listenable: config,
+      builder: (context, child) => SingleChildScrollView(
+        padding: EdgeInsets.all(16),
+        child: Column(
+          crossAxisAlignment: CrossAxisAlignment.start,
+          children: buildConfigurations(context),
+        ),
+      ),
+    );
+  }
+
+  List<Widget> buildConfigurations(BuildContext context);
+}
+
+class AbstractScenarioConfigurator<CONFIG extends AbstractScenarioConfig>
+    extends ScenarioConfigurator<CONFIG> {
+  const AbstractScenarioConfigurator({super.key, required super.config});
 
   @override
   List<Widget> buildConfigurations(BuildContext context) {